Fix auxiliaries and question formatting
The current round of questions now seems to look nice.

I added a QuestionAnswerPair struct thing to hold questions and answers
and render them when necessary. It may prove useful to keep track of
more information (like which dependencies were used in writing the
question) in case we need to later. Maybe it could also be a nice thing
to use when grouping questions/queries.

As for now, QuestionGenerator just uses the rendered question wholesale
as a single string.