Fix minor question deps bug
The question dependency for the subject was not being included in the
cases where we have to recover the subject by looking at the deps out
from the verb (i.e., PP questions). This is fixed.

Also some very minor cleanup to QuestionGenerator and a way to add
tricky sentences to a list for examination in a sanity check run
for questions.